+++ api-builder-plugin-fn-xml-node/Changelog.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.0.8] 2020-09-07
+### Changed
+- updated @axway/api-builder-test-utils from 1.1.0 to 1.1.4
+- updated @axway/api-builder-sdk from 1.0.0 to 1.0.7
+- updated mocha from 7.1.1 to 7.2.0
+
 ## [1.0.7] 2020-09-07
 ### Changed
 - now using the same selector ($.property) syntax as for any other context variable.